# docs/api/combined.openapi.yaml
openapi: 3.0.3
info:
  title: TZW LTD FEMS - Combined API
  version: 1.0.0
  description: |
    Fire Extinguisher Management System REST API with RBAC (ADMIN, INSPECTOR, USER).
    All API routes are accessed through the gateway at `/api`.

    **Swagger UI:** http://localhost:3000/api/docs  
    **OpenAPI JSON:** http://localhost:3000/api/docs/openapi.json

    Per-service docs (health checks only on individual services): auth `:3001/docs`, extinguisher `:3002/docs`, reporting `:3003/docs`, notification `:3004/docs`.
servers:
  - url: http://localhost:3000/api
    description: Local API Gateway
  - url: http://localhost:3000
    description: Gateway root (health)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceKey:
      type: apiKey
      in: header
      name: x-service-key
  parameters:
...
    UuidPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ExtinguisherType:
      type: string
      enum: [WATER, CO2, FOAM, DRY_CHEMICAL]
    ExtinguisherSize:
      type: string
      enum: [LB_1_5, LB_5, LB_9, LB_12]
    ExtinguisherStatus:
      type: string
      enum: [ACTIVE, EXPIRED, UNDER_MAINTENANCE, DECOMMISSIONED]
...
    ExtinguisherRequest:
      type: object
      required: [serialNumber, location, type, size, installationDate, expiryDate]
      properties:
        serialNumber: { type: string, example: FE-2024-001 }
        location: { type: string, example: Building A - Floor 2 }
        type: { $ref: '#/components/schemas/ExtinguisherType' }
        size: { $ref: '#/components/schemas/ExtinguisherSize' }
        installationDate: { type: string, format: date, example: '2024-01-15' }
        expiryDate: { type: string, format: date, example: '2026-01-15' }
        status: { $ref: '#/components/schemas/ExtinguisherStatus' }
    ExtinguisherUpdateRequest:
      type: object
      minProperties: 1
      properties:
        serialNumber: { type: string }
        location: { type: string }
        type: { $ref: '#/components/schemas/ExtinguisherType' }
        size: { $ref: '#/components/schemas/ExtinguisherSize' }
        installationDate: { type: string, format: date }
        expiryDate: { type: string, format: date }
        status: { $ref: '#/components/schemas/ExtinguisherStatus' }
    InspectionScheduleRequest:
      type: object
      required: [extinguisherId, scheduledDate, scheduledTime, assignedInspectorId, assignedInspectorEmail]
      properties:
        extinguisherId: { type: string, format: uuid }
        scheduledDate: { type: string, format: date, example: '2026-06-15' }
        scheduledTime: { type: string, example: '09:00' }
        assignedInspectorId: { type: string, format: uuid }
        assignedInspectorEmail: { type: string, format: email }
        notes: { type: string }
...
security:
  - bearerAuth: []
paths:
...
  /extinguishers:
    get:
      tags: [Extinguishers]
      summary: List all fire extinguishers
      responses:
        '200': { description: List of extinguishers }
    post:
      tags: [Extinguishers]
      summary: Register new fire extinguisher
      description: "Roles: ADMIN"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtinguisherRequest'
      responses:
        '201': { description: Created }
        '403': { description: Insufficient permissions }
  /extinguishers/{id}:
    get:
      tags: [Extinguishers]
      summary: Get extinguisher by ID
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      responses:
        '200': { description: Extinguisher details }
        '404': { description: Not found }
    put:
      tags: [Extinguishers]
      summary: Update extinguisher
      description: "Roles: ADMIN"
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtinguisherUpdateRequest'
      responses:
        '200': { description: Updated }
        '404': { description: Not found }
        '403': { description: Insufficient permissions }
    delete:
      tags: [Extinguishers]
      summary: Delete extinguisher
      description: "Roles: ADMIN"
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      responses:
        '200': { description: Deleted }
        '404': { description: Not found }
        '403': { description: Insufficient permissions }
  /extinguishers/{id}/integrity:
    get:
      tags: [Extinguishers]
      summary: Extinguisher integrity flags
      description: "Roles: ADMIN. Returns active inspection and maintenance constraints before update/delete."
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      responses:
        '200': { description: Integrity summary for the extinguisher }
        '404': { description: Not found }
        '403': { description: Insufficient permissions }
  /extinguishers/{id}/maintenance:
    get:
      tags: [Maintenance]
      summary: Maintenance history for extinguisher
      description: "Roles: ADMIN, INSPECTOR"
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      responses:
        '200': { description: Maintenance logs for the extinguisher }
  /inspections:
    get:
      tags: [Inspections]
      summary: List inspections (filtered by role)
      description: ADMIN sees all; INSPECTOR sees assigned; USER sees own created.
      responses:
        '200': { description: Inspection list }
    post:
      tags: [Inspections]
      summary: Schedule inspection
      description: "Roles: USER. Sends email to assigned inspector."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionScheduleRequest'
      responses:
        '201': { description: Scheduled with email notification }
  /inspections/{id}:
    get:
      tags: [Inspections]
      summary: Get inspection by ID
      description: Role-scoped access. USER own created; INSPECTOR assigned; ADMIN all.
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      responses:
        '200': { description: Inspection details }
        '403': { description: Insufficient permissions }
        '404': { description: Not found }
  /inspections/{id}/complete:
    patch:
      tags: [Inspections]
      summary: Mark inspection complete
      description: "Roles: INSPECTOR (assigned inspections only)"
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      responses:
        '200': { description: Completed }
        '404': { description: Not found }
        '403': { description: Insufficient permissions }
  /inspections/{id}/cancel:
    patch:
      tags: [Inspections]
      summary: Cancel inspection
      description: USER may cancel own; ADMIN/INSPECTOR any.
      parameters:
        - $ref: '#/components/parameters/UuidPath'
      responses:
        '200': { description: Cancelled }
        '403': { description: Insufficient permissions }
...